requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Caching
pandas
matplotlib
seaborn
scipy
faker
